+++ src/main/kotlin/com/github/ekvedaras/classfactoryphpstorm/integration/definitionMethod/inspection/IncorrectPropertyTypeInspectionInDefinitionMethod.kt
@@ -38,6 +38,7 @@ class IncorrectPropertyTypeInspectionInDefinitionMethod : PhpInspection() {
                         definition.typeForDefinition()
                             .classFactoryTargetOrSelf(expression.project)
                             .unwrapClosureValue(expression.project)
+                            .global(expression.project)
                             .types
                 ).isEmpty()) {
                     holder.registerProblem(